+++ src/plugins/source/config.ts
@@ -92,7 +92,7 @@ Config.prototype.sourceEditorNativeOptions = {
 };
 
 Config.prototype.sourceEditorCDNUrlsJS = [
-	'https://cdnjs.cloudflare.com/ajax/libs/ace/1.4.2/ace.js'
+	'https://cdnjs.cloudflare.com/ajax/libs/ace/1.43.3/ace.js'
 ];
 
 Config.prototype.beautifyHTMLCDNUrlsJS = [